+++ src/sections/Citation.vue
@@ -2,11 +2,11 @@
 import { computed, type ComputedRef, onMounted, ref, watch } from 'vue'
 
 import { fetchFakeCitation, formatCitation } from '@/utils/citation'
-import type { DateImprecise, Organisation } from '@/types/app'
+import type { DateImpreciseLabelled, Organisation } from '@/types/app'
 import type { PointOfContact as Contact, Identifier } from '@/types/iso'
 
 import SectionTitle from '@/components/SectionTitle.vue'
-import Freetext from '@/components/Markdown.vue'
+import Markdown from '@/components/Markdown.vue'
 import GuidanceText from '@/components/GuidanceText.vue'
 import Link from '@/components/Link.vue'
 import Button from '@/components/Button.vue'
@@ -30,7 +30,7 @@ const props = defineProps({
     required: true
   },
   dates: {
-    type: Array as () => DateImprecise[],
+    type: Array as () => DateImpreciseLabelled[],
     required: true
   },
   contacts: {
@@ -58,8 +58,8 @@ const getCitation = async () => {
   // citation.value = await fetchCitation(doi.value)
 }
 
-const setFreetextInput = () => {
-  freetextInput.value = citationFormatted.value
+const setMarkdownInput = () => {
+  markdownInput.value = citationFormatted.value
 }
 
 const orgMagic: Organisation = organisationsData.organisations['basMagic']
@@ -74,7 +74,7 @@ const nullIdentifier: Identifier = {
 const citationProseClasses = ['prose-sm']
 
 let citation = ref<string>('')
-let freetextInput = ref<string>('')
+let markdownInput = ref<string>('')
 
 let identifier: ComputedRef<Identifier> = computed(() => {
   /*
@@ -114,9 +114,9 @@ let authors: ComputedRef<string[]> = computed(() => {
 let publishedYear: ComputedRef<string> = computed(() => {
   const publishedDate = props.dates.find((date) => date.label === 'published')
   if (publishedDate) {
-    return String(publishedDate.value.getFullYear())
+    return String(publishedDate.date.js.getFullYear())
   }
-  return '?'
+  return ''
 })
 
 let publisher: ComputedRef<string> = computed(() => {
@@ -158,9 +158,10 @@ watch(
         border-colour-class="border-black bg-neutral-100"
         :prose-classes="citationProseClasses"
         :content="citation"
+        id="citation-preview"
       ></Prose>
       <div class="space-x-2 flex items-center">
-        <Button @click="setFreetextInput">Copy to input</Button>
+        <Button @click="setMarkdownInput">Copy to input</Button>
         <GuidanceText
           >Click to copy this citation (with
           <Link
@@ -170,6 +171,6 @@ watch(
         >
       </div>
     </div>
-    <Freetext v-if="citation" :input="freetextInput" />
+    <Markdown v-if="citation" :input="markdownInput" />
   </SectionBorder>
 </template>